Move screenshot code back to hooks.js
because it is not used anywhere else anymore
