+++ features/step_definitions/hooks.js
@@ -1,6 +1,35 @@
 /* eslint-disable import/no-extraneous-dependencies */
 const { defineSupportCode } = require('cucumber');
-const { takeScreenshot, slugify } = require('../support/util.js');
+const fs = require('fs');
+
+function slugify(text) {
+  return text.toString().toLowerCase()
+    .replace(/\s+/g, '-') // Replace spaces with -
+    .replace(/[^\w-]+/g, '') // Remove all non-word chars
+    .replace(/--+/g, '-') // Replace multiple - with single -
+    .replace(/^-+/, '') // Trim - from start of text
+    .replace(/-+$/, ''); // Trim - from end of text
+}
+
+function writeScreenShot(data, filename) {
+  const stream = fs.createWriteStream(filename);
+  stream.write(new Buffer(data, 'base64'));
+  stream.end();
+}
+
+function takeScreenshot(screnarioSlug, callback) {
+  browser.takeScreenshot().then((screenshotBuffer) => {
+    if (!fs.existsSync(browser.params.screenshotFolder)) {
+      fs.mkdirSync(browser.params.screenshotFolder);
+    }
+    const screenshotPath = `${browser.params.screenshotFolder}/${screnarioSlug}.png`;
+    writeScreenShot(screenshotBuffer, screenshotPath);
+    console.log(`Screenshot saved to ${screenshotPath}`); // eslint-disable-line no-console
+    if (callback) {
+      callback();
+    }
+  });
+}
 
 defineSupportCode(({ After }) => {
   After((scenario, callback) => {
